// src/shared/dataprovider/api/model/assignBrandRequest.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StatusEnum } from './statusEnum';

export interface AssignBrandRequest { 
    AgentId?: number;
    BrandId?: number;
    Status?: StatusEnum;
    CiCommissionRate?: number;
    CoCommissionRate?: number;
    EffectivityDate? : string;
}